requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
fastapi==0.110.2
filelock==3.15.4
fonttools==4.51.0
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.23.2
packaging==24.0
pandas==2.2.1
pillow==10.3.0
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
sympy==1.12.1
threadpoolctl==3.5.0
torch==2.3.1
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uvicorn==0.29.0
